+++ src/index.js
@@ -1,71 +1,8 @@
 import './stylesheets/style.css';
-
-const BASE_URI = 'https://us-central1-js-capstone-backend.cloudfunctions.net/api/';
-
-const parseResponse = (result) => {
-  const resultStr = result.split(':');
-  const [, secondPart] = resultStr;
-  const trimmedText = secondPart.trim();
-  const endIndex = trimmedText.indexOf(' ');
-  return trimmedText.substring(0, endIndex);
-};
-
-const createGame = async (name) => {
-  const ENDPOINT = 'games/';
-  const URI = `${BASE_URI}${ENDPOINT}`;
-  const response = await fetch(URI, {
-    method: 'POST',
-    body: JSON.stringify(name),
-    headers: {
-      'Content-type': 'application/json; Charset=UTF-8',
-    },
-  });
-  return response.json();
-};
-
-const fetchScore = async ({ data, gameId }) => {
-  const GAMES_BASE_URI = `${BASE_URI}games/`;
-  const ENDPOINT = 'scores/';
-  const URI = `${GAMES_BASE_URI}${gameId}/${ENDPOINT}`;
-  const response = await fetch(URI, {
-    method: 'POST',
-    body: JSON.stringify(data),
-    headers: {
-      'Content-type': 'application/json; Charset=UTF-8',
-    },
-  });
-  return response.json();
-};
-
-const refreshScore = async (gameId) => {
-  const GAMES_BASE_URI = `${BASE_URI}games/`;
-  const ENDPOINT = 'scores/';
-  const URI = `${GAMES_BASE_URI}${gameId}/${ENDPOINT}`;
-  const response = await fetch(URI);
-  return response.json();
-};
-
-const createScore = ({ score, user }) => {
-  const li = document.createElement('li');
-  const div = document.createElement('div');
-  const spanName = document.createElement('span');
-  spanName.textContent = user;
-  const spanCol = document.createElement('span');
-  spanCol.textContent = ' : ';
-  const spanScore = document.createElement('span');
-  spanScore.textContent = score;
-  div.append(spanName, spanCol, spanScore);
-  li.append(div);
-  return li;
-};
-
-const updateDomWithScores = (scores) => {
-  const scoreList = document.getElementById('scores-list');
-  scoreList.innerHTML = '';
-  scores.forEach((score) => {
-    scoreList.appendChild(createScore(score));
-  });
-};
+import {
+  parseResponse, fetchScore, refreshScore, createGame,
+} from './utils.js';
+import updateDomWithScores from './dom.js';
 
 window.addEventListener('DOMContentLoaded', async () => {
   const GAMESNAME = {
@@ -78,6 +15,7 @@ window.addEventListener('DOMContentLoaded', async () => {
   const refreshButton = document.getElementById('refresh-score');
   const user = document.getElementById('your-name');
   const score = document.getElementById('your-score');
+  const scoreList = document.getElementById('scores-list');
 
   addButton.addEventListener('click', async (event) => {
     event.preventDefault();
@@ -93,12 +31,13 @@ window.addEventListener('DOMContentLoaded', async () => {
     user.value = '';
     score.value = '';
     const { result } = await refreshScore(gameId);
-    updateDomWithScores(result);
+    updateDomWithScores(scoreList, result);
+    user.focus();
   });
 
   refreshButton.addEventListener('click', async (event) => {
     event.preventDefault();
     const { result } = await refreshScore(gameId);
-    updateDomWithScores(result);
+    updateDomWithScores(scoreList, result);
   });
 });
